+++ packages/italian-verbs-dict/example.js
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-
 const ItalianVerbsDict = require('./dist/verbs.json');
 
 console.log(ItalianVerbsDict['mangiare']);
