+++ packages/italian-verbs-dict/package.json
@@ -2,7 +2,7 @@
   "name": "italian-verbs-dict",
   "version": "1.1.8",
   "description": "Italian verbs dictionnary, based on morph-it",
-  "main": "dist/verbs.json",
+  "main": "dist/index.js",
   "scripts": {
     "clean": "rm -rf dist",
     "test": "nyc --reporter=lcov --reporter=text mocha",
